fix suggest hide, load onfocus
