+++ src/ui/components/Suggest/Suggest.tsx
@@ -1,4 +1,4 @@
-import React, { FC, LegacyRef } from "react";
+import React, { FC, LegacyRef, memo } from "react";
 import { ICollection } from "../../../models/ICollection";
 import Preloader from "../../elements/Preloader/Preloader";
 import s from "./Suggest.module.css";
@@ -10,30 +10,39 @@ import { EmptyTitle } from "../EmptyTitle/EmptyTitle";
 interface SuggestProps {
   movies: ICollection[] | null | undefined;
   suggestRef: null | LegacyRef<HTMLDivElement>;
+  suggestHided: (isVisible: boolean) => void;
 }
 
-export const Suggest: FC<SuggestProps> = ({ movies, suggestRef }) => {
-  const { theme } = useTheme();
+export const Suggest: FC<SuggestProps> = memo(
+  ({ movies, suggestRef, suggestHided }) => {
+    const { theme } = useTheme();
 
-  return (
-    <div
-      className={cn(s.wrapper, {
-        [s.dark]: theme === "dark",
-        [s.light]: theme === "light",
-      })}
-      ref={suggestRef}
-    >
-      {movies ? (
-        movies.length > 0 ? (
-          movies.map((item) => <SuggestItem key={item.id} {...item} />)
+    return (
+      <div
+        className={cn(s.wrapper, {
+          [s.dark]: theme === "dark",
+          [s.light]: theme === "light",
+        })}
+        ref={suggestRef}
+      >
+        {movies ? (
+          movies.length > 0 ? (
+            movies.map((item) => (
+              <SuggestItem
+                key={item.id}
+                {...item}
+                suggestHided={suggestHided}
+              />
+            ))
+          ) : (
+            <EmptyTitle title="Ничего не нашлось" classNames={s.emptyTitle} />
+          )
         ) : (
-          <EmptyTitle title="Ничего не нашлось" classNames={s.emptyTitle} />
-        )
-      ) : (
-        <div className={s.preloader}>
-          <Preloader width={30} />
-        </div>
-      )}
-    </div>
-  );
-};
+          <div className={s.preloader}>
+            <Preloader width={30} />
+          </div>
+        )}
+      </div>
+    );
+  },
+);
